requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
postgres==4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
